Move ownership of Player and Move to Game
