+++ src/board/Board.tsx
@@ -1,4 +1,4 @@
-import { Move } from "../types/Types";
+import { Move } from "../Game";
 import BoardRow from "./BoardRow";
 import LabelRow from "./LabelRow";
