+++ src/board/Cell.tsx
@@ -1,4 +1,4 @@
-import { Move } from "../types/Types";
+import { Move } from "../Game";
 
 interface CellProps {
     className: string;
