--- src/types/Types.ts
@@ -1,2 +1,0 @@
-export type Player = 'X' | 'O';
-export type Move = Player | null;
